Add / to beginning of route names
Previously, routes had worked without the explicit need of a slash in
front of the route. However, in an update, it seems this no longer
works, and instead, the current route is not cleared when a new link is
clicked. Adding a / makes it so that the route is cleared upon clicking
a new Link component.